fix: need to split the user registration function out for __init__.py; correct the package dependency
