+++ requirements_arm.txt
@@ -85,9 +85,9 @@ pyarrow==15.0.0
 pyarrow-hotfix==0.6
 pyclipper==1.3.0.post5
 pycparser==2.21
-pycryptodome==3.20.0
-pycryptodome-test-vectors==1.0.14
-pycryptodomex==3.20.0
+pycryptodome
+pycryptodome-test-vectors
+pycryptodomex
 pydantic==2.6.2
 pydantic_core==2.16.3
 PyJWT==2.8.0
@@ -98,7 +98,7 @@ python-dateutil==2.8.2
 python-docx==1.1.0
 python-dotenv==1.0.1
 python-pptx==0.6.23
-pytz==2024.1
+pytz
 PyYAML==6.0.1
 redis==5.0.3
 regex==2023.12.25
@@ -137,7 +137,7 @@ loguru==0.7.2
 PyJWT==2.8.0
 umap-learn
 fasttext==0.9.2
-volcengine
+volcengine==1.0.141
 opencv-python-headless==4.9.0.80
 readability-lxml==0.8.1
 html_text==0.6.2
