+++ requirements_dev.txt
@@ -72,9 +72,9 @@ pyarrow==15.0.0
 pyarrow-hotfix==0.6
 pyclipper==1.3.0.post5
 pycparser==2.21
-pycryptodome==3.20.0
-pycryptodome-test-vectors==1.0.14
-pycryptodomex==3.20.0
+pycryptodome
+pycryptodome-test-vectors
+pycryptodomex
 pydantic==2.6.2
 pydantic_core==2.16.3
 PyJWT==2.8.0
@@ -85,7 +85,7 @@ python-dateutil==2.8.2
 python-docx==1.1.0
 python-dotenv==1.0.1
 python-pptx==0.6.23
-pytz==2024.1
+pytz
 PyYAML==6.0.1
 regex==2023.12.25
 requests==2.31.0
@@ -125,6 +125,6 @@ redis==5.0.4
 PyJWT==2.8.0
 fasttext==0.9.2
 umap-learn
-volcengine
+volcengine==1.0.141
 readability-lxml==0.8.1
 html_text==0.6.2
